Add missing flags to mimic init functionality
